requirements.txt
fastapi==0.111.0
PyYAML==6.0.1
Requests==2.32.3
# openai-whisper==20231117
# torch==2.3.1
# torchaudio==2.3.1
# torchvision==0.18.1
numpy<2
uvicorn
PyYAML
langchain_cohere